+++ CHANGELOG.md
@@ -1,6 +1,11 @@
-## 0.2.0
+## [0.2.1]
+-  Implemented Binah video service v2
+-  Add `YouthDataPoint` as main biomarker model
+-  Fix Dinah dispose function 
+
+## [0.2.0]
 -  Implemented Binah video service v1
 -  Warning and Error callbacks handled in appropriate format for the client side
 
-## 0.1.0
-- initial version of the biomarkers SDK
+## [0.1.0]
+- Initial version of the biomarkers SDK